Clean up parser (first work for JPEG decoder)
Former-commit-id: 47067e0f2973993cafc99e2b9d000fe42ccb7e96 [formerly ac3f085287cda5eae5c5d923229eb11a867be465] [formerly 7f733bbac0b36cab4a960aed92227e4b1f4488be [formerly 1453b0cbbe11421dc12084aa2db757c3b83c42ff] [formerly cdfa0e12cf1bbc33a13b9c3076b0c7d077469dbb] [formerly 6925c7ee1846c6da3c63ddafc91796b3a888e3a1 [formerly cec58c28c6e9a795eba8111e5b84b62bace4a4b8]] [formerly 7f733bbac0b36cab4a960aed92227e4b1f4488be [formerly 1453b0cbbe11421dc12084aa2db757c3b83c42ff] [formerly cdfa0e12cf1bbc33a13b9c3076b0c7d077469dbb] [formerly 6925c7ee1846c6da3c63ddafc91796b3a888e3a1 [formerly cec58c28c6e9a795eba8111e5b84b62bace4a4b8]] [formerly 5dda768972dbb7d0461ca6a2ba8c11c32856fccb [formerly 774786f7523eba2fe46a53bd946015018697095c]]]]
Former-commit-id: 16aef2f04712b484bf1dca90ef9d76592dc24d76
Former-commit-id: 84b3050ecb10f6ec206a7fa2f11ccfbc414b0694
Former-commit-id: fb3078a17e0628b15a30a94d8dc73043601cd6ef
Former-commit-id: fcf03cc0611c88182d26b1fcdbef608bde2560e6
Former-commit-id: e2906cdf9a9814907b77e4e00b318a01ecf0f743 [formerly 7ca93cee9a9daabf89ca5f5afe7640432d49ae7d]
Former-commit-id: 01a392949c2bd6430bf9553d8eede477a9d7687c
Former-commit-id: 8e7f6bfe7852169d854865300f8beb320713967f

+++ Source/RawParser/Model/Parser/Decoder/RawDecoder.cs
@@ -172,7 +172,7 @@ namespace RawNet
                 if (bitPerPixel > 16)
                     throw new RawDecoderException("readUncompressedRaw: Unsupported bit depth");
 
-                uint skipBits = (uint)(inputPitch - w * rawImage.cpp * bitPerPixel / 8);  // Skip per line
+                int skipBits = (int)(inputPitch - w * rawImage.cpp * bitPerPixel / 8);  // Skip per line
                 if (offset.height > rawImage.raw.dim.height)
                     throw new RawDecoderException("readUncompressedRaw: Invalid y offset");
                 if (offset.width + size.width > rawImage.raw.dim.width)
@@ -199,7 +199,7 @@ namespace RawNet
                         bits.CheckPos();
                         for (uint x = 0; x < w; x++)
                         {
-                            uint b = bits.GetBits((uint)bitPerPixel);
+                            uint b = bits.GetBits(bitPerPixel);
                             rawImage.raw.data[x + (offset.width * rawImage.cpp + y * rawImage.raw.dim.width * rawImage.cpp)] = (ushort)b;
                         }
                         bits.SkipBits(skipBits);
@@ -215,7 +215,7 @@ namespace RawNet
                         bits.CheckPos();
                         for (uint x = 0; x < w; x++)
                         {
-                            uint b = bits.GetBits((uint)bitPerPixel);
+                            uint b = bits.GetBits(bitPerPixel);
                             dest[x] = (ushort)b;
                         }
                         bits.SkipBits(skipBits);
@@ -231,7 +231,7 @@ namespace RawNet
                         bits.CheckPos();
                         for (uint x = 0; x < w; x++)
                         {
-                            uint b = bits.GetBits((uint)bitPerPixel);
+                            uint b = bits.GetBits(bitPerPixel);
                             dest[x] = (ushort)b;
                         }
                         bits.SkipBits(skipBits);
@@ -257,7 +257,7 @@ namespace RawNet
                         bits.CheckPos();
                         for (uint x = 0; x < w; x++)
                         {
-                            uint b = bits.GetBits((uint)bitPerPixel);
+                            uint b = bits.GetBits(bitPerPixel);
                             dest[x] = (ushort)b;
                         }
                         bits.SkipBits(skipBits);
